# apps/web/src/lib/contracts/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hubris-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/chats/{conversation_id}": {
      "get": {
        "tags": ["crate::api::chats"],
        "operationId": "get_chat",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatConversationDetail"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{conversation_id}/interrupt": {
      "post": {
        "tags": ["crate::api::chats"],
        "operationId": "interrupt_chat",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Interrupt requested"
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conversation not running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{conversation_id}/messages": {
      "post": {
        "tags": ["crate::api::chats"],
        "operationId": "send_chat_message",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Message accepted"
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to start Codex runtime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events": {
      "get": {
        "tags": ["crate::api::events"],
        "operationId": "event_stream",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent event stream"
          }
        }
      }
    },
...
    "/api/projects/{project_id}/worktrees/{worktree_id}/chats": {
      "get": {
        "tags": ["crate::api::chats"],
        "operationId": "list_project_worktree_chats",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "worktree_id",
            "in": "path",
            "description": "Worktree ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatConversationSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ChatConversationDetail": {
        "type": "object",
        "description": "Full chat detail payload used to hydrate an open chat tab.",
        "required": ["conversation", "messages"],
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ChatConversationSummary"
          },
          "latestRun": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChatRun"
              }
            ]
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          }
        }
      },
      "ChatConversationSummary": {
        "type": "object",
        "description": "Persisted summary for a conversation list row.",
        "required": [
          "id",
          "sessionId",
          "projectId",
          "worktreeId",
          "provider",
          "title",
          "createdAt",
          "updatedAt",
          "lastActivityAt",
          "lastRunState",
          "revision"
        ],
        "properties": {
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "lastActivityAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "lastError": {
            "type": ["string", "null"]
          },
          "lastMessageAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "minimum": 0
          },
          "lastRunState": {
            "$ref": "#/components/schemas/ChatRunStatus"
          },
          "openTabId": {
            "type": ["string", "null"]
          },
          "projectId": {
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/ChatProvider"
          },
          "providerThreadId": {
            "type": ["string", "null"]
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "sessionId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "worktreeId": {
            "type": "string"
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "description": "Persisted transcript message.",
        "required": [
          "id",
          "conversationId",
          "role",
          "status",
          "contentText",
          "sequence",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "contentText": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "providerTurnId": {
            "type": ["string", "null"]
          },
          "role": {
            "$ref": "#/components/schemas/ChatMessageRole"
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "status": {
            "$ref": "#/components/schemas/ChatMessageStatus"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ChatMessageRole": {
        "type": "string",
        "description": "Persisted message role for a conversation transcript item.",
        "enum": ["user", "assistant"]
      },
      "ChatMessageStatus": {
        "type": "string",
        "description": "Persisted message lifecycle state.",
        "enum": ["pending", "streaming", "completed", "interrupted", "failed"]
      },
      "ChatProvider": {
        "type": "string",
        "description": "Supported chat providers.",
        "enum": ["codex"]
      },
      "ChatRun": {
        "type": "object",
        "description": "Persisted run summary.",
        "required": ["id", "conversationId", "status", "startedAt"],
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "errorMessage": {
            "type": ["string", "null"]
          },
          "finishedAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "providerTurnId": {
            "type": ["string", "null"]
          },
          "startedAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "status": {
            "$ref": "#/components/schemas/ChatRunStatus"
          }
        }
      },
      "ChatRunStatus": {
        "type": "string",
        "description": "Persisted run lifecycle state.",
        "enum": ["starting", "running", "completed", "interrupted", "failed"]
      },
...
      "SendChatMessageRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}